+++ benchmark/fast-find.cpp
@@ -8,7 +8,7 @@
 #include "search_policies.hpp"
 #include "simple_store_policy.hpp"
 
-constexpr int number_of_samples = 20;
+constexpr int number_of_samples = 10;
 constexpr int number_of_operations = 1;
 
 int test(const std::string &command, const std::string &path) {
